ashnova.v2/services/simple_sns_api/requirements.txt
fastapi>=0.110
mangum>=0.17
boto3>=1.34
pydantic>=2.6
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
uvicorn>=0.27
azure-cosmos>=4.7
azure-storage-blob>=12.19
google-cloud-firestore>=2.16
google-cloud-storage>=2.19
azure-functions>=1.20
